requirements.txt
asgiref==3.2.3
django==2.1.5
pytz==2019.3
sqlparse==0.3.0
django-rest-framework==0.1.0
django-rest-swagger==2.1.1
coreapi==2.3.3